pyproject.toml
[project]
name = "fatterbox"
version = "1.0.0"
description = "fatterbox"
requires-python = "==3.11.*"
dependencies = [
    # EXPORT HF_TOKEN="none"
    "chatterbox-tts",
    "wyoming",
    "torch",
    "torchaudio",
    "perth",
]
